requirements.txt
gradio==1.6.4
matplotlib==3.3.4
torch==1.5.0
seaborn==0.11.0
glog==0.3.1
yacs==0.1.7
numpy==1.19.2
lpips_pytorch==latest
kornia==0.3.1
torchvision==0.8.1
Pillow==8.2.0
scikit_learn==0.24.1
